ch03/pyproject.toml
[project]
name = "ch03"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.49.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.10",
    "langchain>=1.0.3",
    "langchain-google-genai>=3.0.0",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.50.0",
]
